+++ bumblebee/inc/actions/groups.php
@@ -47,8 +47,13 @@ class ActionGroups extends ActionAction  {
     if (! isset($this->PD['id'])) {
       $this->select(issetSet($this->PD, 'showdeleted', false));
     } elseif (isset($this->PD['delete'])) {
-      $this->delete();
+      if ($this->readOnly) {
+        $this->readOnlyError();
+      } else {
+        $this->delete();
+      }
     } else {
+      if ($this->readOnly) $this->_dataCleanse('id');
       $this->edit();
     }
     echo "<br /><br /><a href='".makeURL('groups')."'>".T_('Return to group list')."</a>";
