+++ bumblebee/inc/actions/instrumentclass.php
@@ -47,8 +47,13 @@ class ActionInstrumentClass extends ActionAction  {
     if (! isset($this->PD['id'])) {
       $this->select();
     } elseif (isset($this->PD['delete'])) {
-      $this->delete();
+      if ($this->readOnly) {
+        $this->readOnlyError();
+      } else {
+        $this->delete();
+      }
     } else {
+      if ($this->readOnly) $this->_dataCleanse('id');
       $this->edit();
     }
     echo "<br /><br /><a href='".makeURL('instrumentclass')."'>"
