+++ bumblebee/inc/actions/userclass.php
@@ -46,8 +46,13 @@ class ActionUserClass extends ActionAction  {
     if (! isset($this->PD['id'])) {
       $this->select();
     } elseif (isset($this->PD['delete'])) {
-      $this->delete();
+      if ($this->readOnly) {
+        $this->readOnlyError();
+      } else {
+        $this->delete();
+      }
     } else {
+      if ($this->readOnly) $this->_dataCleanse('id');
       $this->edit();
     }
     echo "<br /><br /><a href='".makeURL('userclass')."'>".T_('Return to user class list')."</a>";
